+++ Objective3/Figures/Elipse.cpp
@@ -1,0 +1,3 @@
+﻿#include "Ellipse.h"
+double Ellipse::Perimetr(){ return 4 * (pi * a * b + pow(a - b, 2)) / (a + b); }
+double Ellipse::Area(){ return a * b * pi; }
