--- Objective3/Figures/Figure.cpp
@@ -1,3 +1,0 @@
-﻿#include <cmath>
-#include <iostream>
-#include "Figure.h"
