+++ Objective3/Figures/Rhomb.cpp
@@ -1,0 +1,3 @@
+﻿#include "Rhomb.h"
+double Rhomb::Perimetr(){ return 4 * a; }
+double Rhomb::Area(){ return  a * h; }
